Move builder tests to test subfolder
