+++ builder/BUILD
@@ -42,89 +42,3 @@ cc_library(
     deps = ["//runtime", ":builder"],
     visibility = ["//visibility:public"],
 )
-
-cc_test(
-    name = "checks_test",
-    srcs = ["checks_test.cc"],
-    deps = [
-        ":builder",
-        "@googletest//:gtest_main",
-        "//runtime",
-    ],
-    size="small",
-)
-
-cc_test(
-    name = "copy_test",
-    srcs = ["copy_test.cc"],
-    deps = [
-        ":builder",
-        "@googletest//:gtest_main",
-        "//runtime",
-    ],
-    size="small",
-)
-
-cc_test(
-    name = "elementwise_test",
-    srcs = ["elementwise_test.cc"],
-    deps = [
-        ":builder",
-        "@googletest//:gtest_main",
-        "//runtime",
-    ],
-    size="small",
-)
-
-cc_test(
-    name = "pipeline_test",
-    srcs = [
-        "pipeline_test.cc", 
-        "bazel_util.h",
-        "test_util.h",
-    ],
-    data = ["replica_pipeline_test.cc"] + glob(["visualize/*.html"]),
-    deps = [
-        ":builder",
-        ":replica_pipeline",
-        "@bazel_tools//tools/cpp/runfiles",
-        "@googletest//:gtest_main",
-        "//runtime",
-        "//runtime:thread_pool",
-        "//runtime:visualize",
-    ],
-    size="small",
-)
-
-cc_test(
-    name = "replica_pipeline_test",
-    srcs = ["replica_pipeline_test.cc"],
-    deps = [
-        ":builder",
-        ":replica_pipeline",
-        "@googletest//:gtest_main",
-    ],
-    size="small",
-)
-
-cc_test(
-    name = "simplify_test",
-    srcs = ["simplify_test.cc"],
-    deps = [
-        ":builder",
-        "@googletest//:gtest_main",
-        "//runtime",
-    ],
-    size="small",
-)
-
-cc_test(
-    name = "substitute_test",
-    srcs = ["substitute_test.cc"],
-    deps = [
-        ":builder",
-        "@googletest//:gtest_main",
-        "//runtime",
-    ],
-    size="small",
-)
